pyproject.toml
[tool.poetry]
name = "drf-api"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.3"
python-dotenv = "^1.0.1"
djangorestframework = "^3.15.2"
psycopg2 = "^2.9.10"
pillow = "^11.0.0"
ipython = "^8.29.0"
django-filter = "^24.3"
djangorestframework-simplejwt = "^5.3.1"
coverage = "^7.6.8"
drf-yasg = "^1.21.8"
stripe = "^11.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
